fixup rack thumb colours processing
Make the state code in coloredObject() a mandatory parameter and update
printRackThumbImage() to make sure it is always properly set. This
restores the colour-based highlighting for objects that don't have any
colours associated through assigned tags. Reduce the footprint of the
colour list argument. Prune and rearrange some code for clarity.