fix deleting of tags after commit 756009b
The call to usePreparedDeleteBlade() in destroyTag() [expectedly] started
to trigger an exception. However, that call didn't do any useful work
in the first place because a foreign key does not allow to delete a row
from TagTree if there are any dependent rows in TagStorage, this way if a
tag can be deleted it means it is not assigned to any rack or object and
hence its colour does not make any difference to any rack thumbnail.

Without the call that truncates RackThumbnail destroyTag() has no purpose
anymore so delete it and leave the $opspec_list declaration to do the job
through tableHandler() as before.