formally retire support for PHP older than 5.5
RackTables used to require PHP 5.2.10 or later for several years. As soon
as PHP versions 7.x are the default now it is wrong to claim support for
something that the upstream itself does not support anymore (5.5 is the
oldest supported version on php.net). This upcoming change was announced
on racktables-users on 2017-07-21 and nobody has objected so let's go
ahead.

From now on it is OK to _improve_ RackTables PHP code in ways that break
it on PHP versions 5.2/5.3/5.4 only.