change some L2 address format processing
Make the L2 address regexps strictly uppercase because they are used to
match all-uppercase strings.

In l2addressForDatabase() add handling of IP-over-InfiniBand L2 addresses
(20 bytes long). Change the SunOS-specific notation to apply to MAC
addresses only. It used to apply to WWN addresses too, most likely due to
a copy-and-paste change, and can be restored if there is a genuine use
case for it.

Resize the text inputs in renderPortsForObject() for IPoIB L2 addresses.

Change l2addressFromDatabase() to reject addresses that it would be
impossible to have in the database by normal means.

Rearrange some code for readability and update comments. Add and update
tests as necessary.