split getObjectClass() into three functions
getObjectClass() used to do three different jobs and to accept two to
three arguments, of which the first was always a constant. Split its code
into three smaller and better focused functions with less arguments and
replace a global with a static. Also in the gradient definition convert
the percentage from float to integer to avoid potential issues in future
(see commit c3fe089).

* getObjectClass(): unary or binary function
* getTagClass(): unary function
* getCachedCSSClassForStyle(): binary function
* renderTagRowForEditor(): update as necessary
* getColorSelect(): ditto
* getTagClassName(): ditto
* enableTagsPicker(): ditto
* setEntityColors(): ditto
* printObjectDetailsForRenderRack(): ditto
* renderRack(): ditto
* renderDepot(): ditto
* renderIPSpaceRecords(): ditto