tests: simplify RenderDepotTest
Test that renderDepot() has not returned an empty string, it is not much
but more than the mere fact the function managed to return without an
exception. Leave output buffer management to getOutputOf() and exception
handling -- to PHPUnit.

This change makes it possible to see that this test failure on PHP 7.1 is
actually an error (as much as PHPUnit distinguishes "failures" and
"errors") and it occurs the same way on both PHPUnit 5 and PHPUnit 6:

There was 1 error:

1) RenderDepotTest::testRenderDepot
session_start(): Cannot send session cookie - headers already sent by
(output started at /usr/share/php/PHPUnit6/Util/Printer.php:112)