dismiss a few unnecessary calls to intval()
In renderVLANDomainListEditor() $dominfo['group_id'] is either NULL or a
string representation of an unsigned integer. Both kinds of values will
evaluate without the conversion as expected in printNiftySelect() and its
worker functions.

In renderEmptyResults() and getPageNumOfIPv6() the value originates from
getConfigVar(), which makes the conversion because the configuration
variable is declared "uint", hence the call is a no-op.