always enable the strict SQL mode
When the strict SQL mode is enabled, the server will fail a query that
tries to INSERT or UPDATE a row that has missing or invalid data. When
it is disabled, the server will try to substitute 0 or an empty string
instead of failing the query.

Historically, the strict SQL mode did not exist in MySQL when RackTables
was first developed, and when it became available in later MySQL versions
it was not enabled by default, so RackTables code didn't account for its
effects for a long time. When the strict SQL mode became the default in
later MySQL versions, RackTables SQL queries were fixed to comply with
it, though it was presumed that the strict SQL mode was not a dependency.

However, it turns out a few SQL triggers replace a semantically invalid
value with a NULL (with the column defined NOT NULL) to make the query
fail --- this behaviour actually depends on the strict SQL mode.

In particular, Fedora 26 comes with MariaDB 10.1.26, which does not have
the strict mode enabled by default (the documentation explains it had
changed in MariaDB 10.2.4). Because of this several unit tests failed
as the invalid data input didn't result in the expected exception class
or any exceptions at all:

There were 6 failures:

1) EntityLinkTriggerTest::testLinkObjectToSelfByUpdate
Failed asserting that exception of type "PDOException" is thrown.

2) EntityLinkTriggerTest::testUpdateLinkBetweenIncompatibleObjects
Failed asserting that exception of type "PDOException" is thrown.

3) EntityLinkTriggerTest::testInvalidateRowLink
Failed asserting that exception of type "PDOException" is thrown.

4) EntityLinkTriggerTest::testInvalidateRackLink
Failed asserting that exception of type "PDOException" is thrown.

5) LinkTriggerTest::testUpdateLinkToSelf
Failed asserting that exception of type "RTDatabaseError" matches
expected exception "PDOException". Message was: "foreign key violation"
at
/usr/local/racktables/wwwroot/inc/database.php:3914
/usr/local/racktables/wwwroot/inc/database.php:4065
/usr/local/racktables/tests/LinkTriggerTest.php:87

6) LinkTriggerTest::testUpdateLinkBetweenIncompatiblePorts
Failed asserting that exception of type "RTDatabaseError" matches
expected exception "PDOException". Message was: "foreign key violation"
at
/usr/local/racktables/wwwroot/inc/database.php:3914
/usr/local/racktables/wwwroot/inc/database.php:4065
/usr/local/racktables/tests/LinkTriggerTest.php:161

As RackTables is already compatible with the strict SQL mode, the best
solution seems to be to enable it during the connection setup. However,
it turns out MariaDB defines STRICT_TRANS_TABLES so that for a
multiple-row query that has invalid or missing data not in the first row
MariaDB will disregard the strict SQL mode and try to convert those
values without failing the query (which is different from MySQL behaviour
and RackTables code assumptions). This way STRICT_ALL_TABLES remains the
only choice as both servers define its effect identically. The fact it
will try to apply to non-transactional tables as well should not be a
problem because all RackTables SQL tables use InnoDB.