tests: request stderr output in phpunit.xml
The change in commit a86d900 fixed RenderDepotTest for express.sh but
standalone "phpunit --group small" would still fail the test for the same
reason as before. To avoid the unnecessary dependency on the command-line
argument move the "stderr" parameter to phpunit.xml (this seems to be
supported in PHPUnit 4.1.0 and later versions).