require a name for a SELECT element
There is no point in having a form and a SELECT with no name in it, thus
when the name attribute is not provided to the function that prints a
SELECT that is a bug. Modify getSelect() and getNiftySelect() to throw
instead of just returning and empty string and hiding the bug.