fix a bug in rebuildTagChainForEntity()
Since my commit 1c09455 it was impossible to modify and save the tag
chain for an unmounted object as it would end up in an exception because
the list of rack IDs was empty. This reproduced at both the "Tags" and
the "Properties" tabs. Put it right: only try to delete thumbnails when
there are any racks associated with the object. Update the comments while
at it.