improve Percona server support (Mantis#1793)
According to the bug report, Atom and ObjectHistory were the only tables
that did not have a unique/primary key. Add an auto_increment primary
key column to ObjectHistory and a unique key to Atom. In the latter case
this converts all columns to NOT NULL and makes the values mandatory due
to the strict SQL mode.