make isUnsignedInteger() unary
This gets rid of the default value and makes the code easier to read. The
function now treats zero is a valid value and any other function that
calls it expects exactly that. Remove the explicit 2nd argument from the
following functions:

* setConfigVar()
* setUserConfigVar()
* HTMLColorFromDatabase()
* sortTokenize()
* constructIPRange()
* updVSTRule()

Add two unary functions -- assertUnsignedIntArg() and
assertNaturalNumArg() -- and reduce assertUIntArg() to a wrapper.

Update the tests.