make the configuration defaults data shareable
Move the array from resetUIConfig() to a new function,
getConfigDefaults(). Place the new function in dictionary.php as that is
the best way to make it available in the installer (which still does not
use it though). Update renderUIResetForm() to use a new opcode that
works better with the hardcoded require_once in index.php.