refine platform tests further
In platform_is_ok() convert the PHP extension tests to declarations and
describe the impact of a failed test iff the test is not critical for the
core RackTables features to work in the default configuration. Ibid,
display the impact of an optional test in the output table when it fails
(this did not work before because platform_function_test() ignored one of
its parameters). Ibid, for the test just use extension_loaded(). Remove
platform_function_test(), which now has no purpose.

[skip ci]