git-commit: make the actions more fine-grained
Introduce an optional -o ("only") option to request only a specific
step of the pull-commit-push workflow. This is useful for updating many
files in a batch -- any changes will still end up one commit per file,
but the workflow will be pull-commit-commit-...-commit-commit-push, which
will not multiply the delay accessing a slow git remote before and after
every attempted file.

Without the option the script will work same as before this change.

[skip ci]