tests: disregard PHPUnit 8 for now
In Travis CI PHP 7.2 now comes with PHPUnit 8, which seems to require
rewriting the tests in a way that is not compatible with older PHPUnit
versions. Introduce SKIP_PHPUNIT into the Travis CI matrix to avoid
running PHPUnit 8 on PHP 7.2 until RackTables code is ready (and
potentially to prevent running older PHPUnit versions on older PHP
versions if the switch happens).