fix hook integration in Cacti and Munin pligins
Commit 011df35 introduced this issue. With "module=image" and either of
the plugins enabled RackTables failed to fail when "img" was set to an
invalid value. This was because the plugins' extension functions always
return TRUE to dispatchImageRequest(), which concluded the work was
always done and thus never raised the default exception. Make the return
in the plugins conditional to fix that.