use vertical writing mode in the 802.1Q report
It used to be inconvenient to view the domain/VLAN matrix on setups with
many VLAN domains. I tried to switch the domain THs to vertical mode
about 7 years ago, but it didn't work well at that time. Nowadays it is
just a matter of a CSS property, it turns out.

* pi.css
* render8021QReport()

[skip ci]