Travis CI: fix the build after Ubuntu upgrade
Travis CI had changed the default Ubuntu version from 14.04 to 16.04,
which had broken the following jobs:
* PHP 5.5 (because Travis does not supply a PHP 5.5 phpenv archive for
  Ubuntu 16.04)
* PHP 7.0 (because PHPUnit was upgraded from 6.4.3 to 7.5.0)
* PHP 7.1 (because PHPUnit was upgraded from 6.4.3 to 8.0.4)

Complete the change started in commit b3a99c9 and specify the latest
PHPUnit 6 for the PHP 7.0 job and the latest PHPUnit 7 for the PHP 7.1
job. Bump PHP 7.2 and 7.3 jobs to the latest PHPUnit 7 too. Remove PHP
5.5 from the build matrix.