Travis CI: use Ubuntu 18.04 and PHP 7.[123]
At the time of writing PHP supports only the following versions:
* 7.1 (old stable)
* 7.2 (old stable)
* 7.3 (current stable)

This list is likely to change once PHP 7.4 is available, which seems to
be coming.

Also PHPUnit supports only the following versions:
* 7 (soon to be retired), which requires PHP 7.1
* 8, which requires PHP 7.2+

In Travis CI Ubuntu 18.04 comes with PHPUnit 8 by default, and supports
only PHP versions 7.1, 7.2 and 7.3. Switch the master branch to 18.04 to
make it more straightforward to rewrite (see commit f81030a) the tests
for PHPUnit 8.

Leave the maintenance-0.21.x branch on Ubuntu 16.04 with its unsupported
versions of PHP and PHPUnit.