add PHPUnit results cache to .gitignore
The results cache was introduced in PHPUnit 7.3 and became on by default
in PHPUnit 8. The feature makes no difference in a [single-use] Travis CI
environment, but can enable defects-first test order in a developer
working copy, in case anybody needs it. Hence let's keep it enabled and
ignore the cache file.

[skip ci]