tests: remove assertions that broke on PHP >= 8.0
As https://www.php.net/manual/en/migration80.incompatible.php documents
it, each of the following expressions evaluates as true in PHP before
8.0 and evaluates as false in PHP 8.0 and later:

0 == ''
'abc' == 0
'' == 0

This made three data sets fail in PureFunctionTest::testUnarySame():

('nullIfEmptyStr', 0, null)
('nullIfZero', 'abc', null)
('nullIfZero', '', null)

For the purposes of RackTables PHP code there is little point in testing
nullIfEmptyStr(), which takes a string argument, on 0 (an integer), and
in testing nullIfZero(), which takes an integer, on not-an-integer
strings.  This is neither the expected nor the implemented behaviour, if
an evidence of the opposite emerges, it should be treated as a bug.
Remove the data points.