manage database mutexes in a more portable way
It is not immediately obvious when PDOStatement::fetchColumn() started
to return SQL integers as PHP integers, but on Debian 12 (PHP 8.2.7,
MariaDB 10.11.4) this is certainly the case.  This at some point had
broken both setDBMutex() and releaseDBMutex(), which used to compare the
result strictly with a string.  The unit test made it obvious:

1) DBMutexTest::testExisting
Failed asserting that false is identical to true.

Use loose comparison in both functions, which should work correctly with
both the old and the new PDO behaviour.